Fix clobbering of string owned by Python
It turns out that Python is pretty smart and gives us a borrowed pointer to an internal string when we extract a C string from the method argument. In this way, when we were chopping up the prefix string argument, we were messing with Python strings. Also, it seems that Python is smart enough to use the same storage for two identical strings, so when we used the same test string twice, Python didn't bother recreating it, and instead tried to use our now-borked string. Cool!